backend/requirements.txt
# FastAPI and dependencies
fastapi==0.116.1
uvicorn[standard]==0.24.0

# Database
motor==3.1.2  # Async MongoDB driver
pymongo==4.3.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6

# WebSocket
websockets==12.0

# Security
python-dotenv==1.0.0
cors==1.0.1

# File handling
aiofiles==23.2.1
Pillow==10.1.0

# Video processing (FFMPEG)
ffmpeg-python==0.2.0

# Validation
pydantic==2.5.0
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.27.0
certifi==2024.2.2
